+++ lib/provider/hive/preferences.dart
@@ -14,12 +14,12 @@
 // along with this program. If not, see
 // <https://www.gnu.org/licenses/agpl-3.0.html>.
 
+import 'package:flutter/material.dart';
 import 'package:hive_flutter/adapters.dart';
 
 import '/store/model/preferences.dart';
 import 'base.dart';
 
-// TODO: Encrypt stored data.
 /// [Hive] storage for a [PreferencesData].
 class PreferencesHiveProvider extends HiveBaseProvider<PreferencesData> {
   @override
@@ -38,6 +38,14 @@ class PreferencesHiveProvider extends HiveBaseProvider<PreferencesData> {
   WindowPreferences? getWindowPreferences() => getSafe(0)?.windowPreferences;
 
   /// Stores a new [WindowPreferences] to [Hive].
-  Future<void> setWindowPreferences(WindowPreferences prefs) =>
-      putSafe(0, (box.get(0) ?? PreferencesData())..windowPreferences = prefs);
+  Future<void> setWindowPreferences({Size? size, Offset? position}) => putSafe(
+        0,
+        (box.get(0) ?? PreferencesData())
+          ..windowPreferences = WindowPreferences(
+            width: size?.width,
+            height: size?.height,
+            dx: position?.dx,
+            dy: position?.dy,
+          ),
+      );
 }